Update LatLong_SchsacRegionChbCounty.csv
Lake of the Woods county is a part of North County CHB. If the map is drawn in ggplot, the lines all lined up before even if Lake of the Woods did not have a chbOrder. However, when building the map in plotly or in ggirafe, these objects did not appreciate that Lake of the Woods was missing the order. This resulted in me making Lake of the Woods its own group for a CHB and ordering it as its own group. This should work since the label and mapping will occur with respect to the CHB and since the geographies are not connected, having Lake of the Woods as its own group should fix issues with Plotly and GGgirafe.